Add support for removing configs

+++ git_ssh/config/remove.py
@@ -39,12 +39,12 @@ class RemoveConfig:
 
     def remove(self):
         """Remove this config"""
-        return self._source.remove()
+        return self._source.remove(self._name)
 
 
 class EmptyRemoveSource:
 
-    def remove(self):
+    def remove(self, name):
         return False
 
 
@@ -57,12 +57,18 @@ class RemoveSource:
         """
         self._path = path
 
-    def remove(self):
+    def remove(self, name):
         """Remove the file location at _path from the filesystem
 
         If the path does not point to a valid file, False is returned
         If the os fails to remove the path, False is returned
         """
+        if not self._path:
+            raise RuntimeError("Cannot call remove() with invalid path")
+
+        if not name:
+            raise RuntimeError("Cannot call remove() with invalid name")
+
         try:
             os.remove(self._path)
         except Exception as e:
@@ -70,4 +76,5 @@ class RemoveSource:
             Logger.e(e)
             return False
         else:
+            Logger.log("Config removed: {} at {}".format(name, self._path))
             return True
